+++ formatter/table.go
@@ -2,7 +2,6 @@ package formatter
 
 import (
 	"github.com/olekukonko/tablewriter"
-	"strconv"
 	"strings"
 	"wander/nomad"
 )
@@ -40,7 +39,7 @@ func createTableConfig(numCols int) tableConfig {
 	return tableConfig{table, tableString}
 }
 
-func getRenderedTableString(columnTitles []string, data [][]string) Table {
+func GetRenderedTableAsString(columnTitles []string, data [][]string) Table {
 	table := createTableConfig(len(columnTitles))
 	table.writer.SetHeader(columnTitles)
 	table.writer.AppendBulk(data)
@@ -51,25 +50,6 @@ func getRenderedTableString(columnTitles []string, data [][]string) Table {
 	return Table{headerRows, contentRows}
 }
 
-func JobResponsesAsTable(jobResponse []nomad.JobResponseEntry) Table {
-	var jobResponseRows [][]string
-	for _, row := range jobResponse {
-		jobResponseRows = append(jobResponseRows, []string{
-			row.ID,
-			row.Type,
-			row.Namespace,
-			strconv.Itoa(row.Priority),
-			row.Status,
-			formatTimeNs(row.SubmitTime),
-		})
-	}
-
-	return getRenderedTableString(
-		[]string{"ID", "Type", "Namespace", "Priority", "Status", "Submit Time"},
-		jobResponseRows,
-	)
-}
-
 func AllocationsAsTable(allocations []nomad.AllocationRowEntry) Table {
 	var allocationResponseRows [][]string
 	for _, alloc := range allocations {
@@ -78,12 +58,12 @@ func AllocationsAsTable(allocations []nomad.AllocationRowEntry) Table {
 			alloc.Name,
 			alloc.TaskName,
 			alloc.State,
-			formatTime(alloc.StartedAt),
-			formatTime(alloc.FinishedAt),
+			FormatTime(alloc.StartedAt),
+			FormatTime(alloc.FinishedAt),
 		})
 	}
 
-	return getRenderedTableString(
+	return GetRenderedTableAsString(
 		[]string{"Alloc ID", "Alloc Name", "Task Name", "State", "Started", "Finished"},
 		allocationResponseRows,
 	)
@@ -100,7 +80,7 @@ func LogsAsTable(logs []nomad.LogRow, logType nomad.LogType) Table {
 		}
 	}
 
-	return getRenderedTableString(
+	return GetRenderedTableAsString(
 		[]string{logType.String()},
 		logRows,
 	)
