+++ nomad/models.go
@@ -5,51 +5,6 @@ import (
 	"time"
 )
 
-type JobResponseEntry struct {
-	ID                string      `json:"ID"`
-	ParentID          string      `json:"ParentID"`
-	Name              string      `json:"Name"`
-	Namespace         string      `json:"Namespace"`
-	Datacenters       []string    `json:"Datacenters"`
-	Multiregion       interface{} `json:"Multiregion"`
-	Type              string      `json:"Type"`
-	Priority          int         `json:"Priority"`
-	Periodic          bool        `json:"Periodic"`
-	ParameterizedJob  bool        `json:"ParameterizedJob"`
-	Stop              bool        `json:"Stop"`
-	Status            string      `json:"Status"`
-	StatusDescription string      `json:"StatusDescription"`
-	JobSummary        struct {
-		JobID     string `json:"JobID"`
-		Namespace string `json:"Namespace"`
-		Summary   struct {
-			YourProjectName struct {
-				Queued   int `json:"Queued"`
-				Complete int `json:"Complete"`
-				Failed   int `json:"Failed"`
-				Running  int `json:"Running"`
-				Starting int `json:"Starting"`
-				Lost     int `json:"Lost"`
-			} `json:"your_project_name"`
-		} `json:"Summary"`
-		Children struct {
-			Pending int `json:"Pending"`
-			Running int `json:"Running"`
-			Dead    int `json:"Dead"`
-		} `json:"Children"`
-		CreateIndex int `json:"CreateIndex"`
-		ModifyIndex int `json:"ModifyIndex"`
-	} `json:"JobSummary"`
-	CreateIndex    int   `json:"CreateIndex"`
-	ModifyIndex    int   `json:"ModifyIndex"`
-	JobModifyIndex int   `json:"JobModifyIndex"`
-	SubmitTime     int64 `json:"SubmitTime"`
-}
-
-func (e JobResponseEntry) MatchesFilter(filter string) bool {
-	return strings.Contains(e.ID, filter)
-}
-
 // AllocationResponseEntry is returned from GET /v1/job/:job_id/allocations
 // https://www.nomadproject.io/api-docs/jobs#list-job-allocations
 type AllocationResponseEntry struct {
